Expose auditing into history model
